requirements.txt
Pillow==10.1.0
pytesseract==0.3.10